+++ modules/anime/src/main/kotlin/com/thekeeperofpie/artistalleydatabase/anime/list/AnimeUserListViewModel.kt
@@ -23,6 +23,7 @@ import com.thekeeperofpie.artistalleydatabase.anime.media.MediaUtils
 import com.thekeeperofpie.artistalleydatabase.anime.media.UserMediaListController
 import com.thekeeperofpie.artistalleydatabase.anime.media.filter.AnimeSortFilterController
 import com.thekeeperofpie.artistalleydatabase.anime.media.filter.MangaSortFilterController
+import com.thekeeperofpie.artistalleydatabase.anime.media.filter.MediaGenresController
 import com.thekeeperofpie.artistalleydatabase.anime.media.filter.MediaLicensorsController
 import com.thekeeperofpie.artistalleydatabase.anime.media.filter.MediaSortFilterController
 import com.thekeeperofpie.artistalleydatabase.anime.media.filter.MediaTagsController
@@ -55,6 +56,7 @@ class AnimeUserListViewModel @Inject constructor(
     private val settings: AnimeSettings,
     val ignoreList: AnimeMediaIgnoreList,
     private val mediaTagsController: MediaTagsController,
+    private val mediaGenresController: MediaGenresController,
     private val mediaLicensorsController: MediaLicensorsController,
     private val userMediaListController: UserMediaListController,
     private val featureOverrideProvider: FeatureOverrideProvider,
@@ -90,6 +92,7 @@ class AnimeUserListViewModel @Inject constructor(
                 settings = settings,
                 featureOverrideProvider = featureOverrideProvider,
                 mediaTagsController = mediaTagsController,
+                mediaGenresController = mediaGenresController,
                 mediaLicensorsController = mediaLicensorsController,
             ).apply {
                 initialize(
@@ -109,6 +112,7 @@ class AnimeUserListViewModel @Inject constructor(
                 settings = settings,
                 featureOverrideProvider = featureOverrideProvider,
                 mediaTagsController = mediaTagsController,
+                mediaGenresController = mediaGenresController,
                 mediaLicensorsController = mediaLicensorsController,
             ).apply {
                 initialize(