+++ modules/anime/src/main/kotlin/com/thekeeperofpie/artistalleydatabase/anime/media/filter/MediaTagsController.kt
@@ -1,34 +1,35 @@
 package com.thekeeperofpie.artistalleydatabase.anime.media.filter
 
+import android.os.SystemClock
 import com.anilist.MediaTagsQuery
 import com.thekeeperofpie.artistalleydatabase.android_utils.ScopedApplication
 import com.thekeeperofpie.artistalleydatabase.android_utils.kotlin.CustomDispatchers
 import com.thekeeperofpie.artistalleydatabase.anilist.oauth.AuthedAniListApi
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.catch
-import kotlinx.coroutines.flow.collectLatest
+import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.mapLatest
-import kotlinx.coroutines.launch
+import kotlinx.coroutines.flow.shareIn
 
 @OptIn(ExperimentalCoroutinesApi::class)
 class MediaTagsController(scopedApplication: ScopedApplication, aniListApi: AuthedAniListApi) {
 
-    val tags = MutableStateFlow(emptyMap<String, TagSection>())
+    private val refresh = MutableStateFlow(-1L)
 
-    val refreshUptimeMillis = MutableStateFlow(-1L)
+    val tags = refresh.mapLatest {
+        aniListApi.tags().mediaTagCollection
+            ?.filterNotNull()
+            ?.let(::buildTagSections)
+            .orEmpty()
+    }
+        .catch { emit(emptyMap()) }
+        .flowOn(CustomDispatchers.IO)
+        .shareIn(scopedApplication.scope, SharingStarted.Lazily, replay = 1)
 
-    init {
-        scopedApplication.scope.launch(CustomDispatchers.IO) {
-            refreshUptimeMillis.mapLatest {
-                aniListApi.tags().mediaTagCollection
-                    ?.filterNotNull()
-                    ?.let(::buildTagSections)
-                    .orEmpty()
-            }
-                .catch { emit(emptyMap()) }
-                .collectLatest(tags::emit)
-        }
+    fun refresh() {
+        refresh.value = SystemClock.uptimeMillis()
     }
 
     /**